requirements.txt
appnope==0.1.3
asttokens==2.1.0
autopep8==2.0.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
jupyter_client==7.4.4
jupyter_core==4.11.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.32
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
six==1.16.0
stack-data==0.6.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
